+++ packages/dbx-core/src/lib/action/directive/state/index.ts
@@ -5,3 +5,5 @@ export * from './action.success.directive';
 export * from './action.value.directive';
 export * from './action.working.component';
 export * from './action.enforce.modified.directive';
+export * from './action.value.trigger.directive';
+export * from './action.value.trigger.instance';