+++ packages/dbx-web/src/lib/interaction/dialog/dialog.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { DbxDialogContentDirective } from './dialog.content.component';
 import { DbxStyleLayoutModule } from '../../layout/style/style.layout.module';
+import { DbxActionDialogDirective } from './dialog.action.directive';
 
 /**
  * Module for block components.
@@ -13,9 +14,11 @@ import { DbxStyleLayoutModule } from '../../layout/style/style.layout.module';
   ],
   declarations: [
     DbxDialogContentDirective,
+    DbxActionDialogDirective
   ],
   exports: [
-    DbxDialogContentDirective
+    DbxDialogContentDirective,
+    DbxActionDialogDirective
   ]
 })
 export class DbxDialogInteractionModule { }