+++ packages/dbx-web/src/lib/interaction/dialog/index.ts
@@ -1,3 +1,4 @@
 export * from './abstract.dialog.directive';
 export * from './dialog.content.component';
+export * from './dialog.action.directive';
 export * from './dialog.module';