+++ packages/demo/src/app/modules/doc/modules/action/component/action.example.tool.component.ts
@@ -1,6 +1,6 @@
 import { filterMaybe } from '@dereekb/rxjs';
 import { BehaviorSubject, shareReplay, switchMap } from 'rxjs';
-import { Component, Input, OnInit, OnDestroy, Optional, Host } from '@angular/core';
+import { Component, Input, OnDestroy, Optional, Host } from '@angular/core';
 import { ProvideFormlyContext } from '@dereekb/dbx-form';
 import { DbxActionContextStoreSourceInstance, DbxActionDirective } from '@dereekb/dbx-core';
 import { Maybe } from '@dereekb/util';