+++ packages/demo/src/app/modules/doc/modules/action/container/interaction.component.ts
@@ -1,7 +1,11 @@
-import { DbxActionSnackbarGeneratorUndoInput } from '@dereekb/dbx-web';
+import { MatDialog } from '@angular/material/dialog';
+import { DbxActionDialogFunction, DbxPopoverService } from '@dereekb/dbx-web';
+import { DbxActionPopoverFunction, DbxActionSnackbarGeneratorUndoInput } from '@dereekb/dbx-web';
 import { ChangeDetectorRef, Component } from '@angular/core';
 import { DbxActionContextMachine, DbxActionContextSourceReference, HandleActionFunction, WorkHandlerContext, safeDetectChanges } from '@dereekb/dbx-core';
 import { of, delay, BehaviorSubject, tap } from 'rxjs';
+import { DocActionExamplePopoverComponent } from '../component/action.example.popover.form.component';
+import { DocActionExampleDialogComponent } from '../component/action.example.dialog.component';
 
 @Component({
   templateUrl: './interaction.component.html'
@@ -14,7 +18,10 @@ export class DocActionInteractionComponent {
   private _value = new BehaviorSubject<{ test: number }>({ test: 0 });
   readonly value$ = this._value.asObservable();
 
-  constructor(readonly cdRef: ChangeDetectorRef) { }
+  constructor(
+    readonly dbxPopoverService: DbxPopoverService,
+    readonly matDialog: MatDialog,
+    readonly cdRef: ChangeDetectorRef) { }
 
   readonly handleAction: HandleActionFunction = (value: any, context: WorkHandlerContext) => {
     return of(true).pipe(delay(1000));
@@ -40,6 +47,14 @@ export class DocActionInteractionComponent {
     });
 
     return instance;
+  }
+
+  handleOpenPopover: DbxActionPopoverFunction = ({ origin }) => {
+    return DocActionExamplePopoverComponent.openPopover(this.dbxPopoverService, { origin });
+  }
+
+  handleOpenDialog: DbxActionDialogFunction = () => {
+    return DocActionExampleDialogComponent.openDialog(this.matDialog);
   }
 
 }