+++ packages/demo/src/app/modules/doc/modules/action/doc.action.module.ts
@@ -11,6 +11,8 @@ import { DocActionContextComponent } from './container/context.component';
 import { DocActionExampleToolsComponent } from './component/action.example.tool.component';
 import { DocActionFormExampleFormComponent } from './component/action.example.form.component';
 import { DocActionMapComponent } from './container/map.component';
+import { DocActionExamplePopoverComponent } from './component/action.example.popover.form.component';
+import { DocActionExampleDialogComponent } from './component/action.example.dialog.component';
 
 @NgModule({
   imports: [
@@ -23,6 +25,8 @@ import { DocActionMapComponent } from './container/map.component';
     // component
     DocActionExampleToolsComponent,
     DocActionFormExampleFormComponent,
+    DocActionExampleDialogComponent,
+    DocActionExamplePopoverComponent,
     // container
     DocActionLayoutComponent,
     DocActionHomeComponent,