+++ packages/demo/src/app/modules/doc/modules/interaction/container/dialog.component.ts
@@ -11,7 +11,6 @@ export class DocInteractionDialogComponent {
 
   openDialog() {
     this.matDialog.open(DocInteractionExampleDialogComponent);
-    
   }
 
 }